Refactor everything out of a single file
