+++ tap_shopify/streams/__init__.py
@@ -1,0 +1,3 @@
+import tap_shopify.streams.orders
+import tap_shopify.streams.metafields
+import tap_shopify.streams.transactions
